+++ routes/web.php
@@ -33,6 +33,7 @@ Route::get('/home', 'HomeController@index')->name('home');
 Route::get('/dashboard', 'Admin\DashboardController@index') -> name('dashboard') -> middleware('can:admin');
 
 Route::get('/search', 'Bible\SearchBibleController@index' );
+Route::get('/verse', 'Bible\SearchVerseController@index');
 
 Auth::routes();